c/charge: add limit support, unused kWh calc
Added --limit/-l option to the `charge start` command to set a charge
limit (by percent).

Also wrote some functions for parsing option codes to determine battery
capacity in kWh, but it turns out none of that is going to work: we
don't actually know what most of the option codes mean. For example, my
Model S with a 100 kWh battery has option codes that seem to indicate it
is a Model 3 with a 75 kWh battery (it definitely isn't). Thus, these
functions are not actually used (yet!).