cmd: add software command(s)
New software commands:

* `software version`: display the version of the firmware currently
running on the vehicle. If an update is available, scheduled, or being
installed, that is shown here, along with estimated installation time.
* `software schedule`: schedule software update installation. Note that
the `--delay/-d seconds` parameter is currently not working due to
limitations either in the API or in the library used here
(http://github.com/kabaka/tesla) – all updates are scheduled 120 seconds
(2 minutes) in the future. Note that this only works if an update is
already available and you have been prompted for installation – this
command cannot trigger or force a new update.
* `software cancel`: cancel the scheduled software update. This can only
be done before installation begins.